test: :white_check_mark: Migracion con exito, sólo unos pequeños cambios en algunas entidades, pero todo nais :D

+++ Core/Entities/Auditoria.cs
@@ -9,7 +9,7 @@ namespace Core.Entities
     public class Auditoria : BaseEntity
     {
         public string NombreUsuario { get; set; }
-        public BigInteger DescAccion { get; set; }
+        public int DescAccion { get; set; }
         public DateTime FechaCreacion { get; set; }
         public DateTime FechaModificacion { get; set; }
         public ICollection<BlockChain> BlockChains { get; set; }
